+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - The skip action will now listen to the window events instead of the player element by default
 - Added `isNarrator` to the dialog state
 
+### Fixed
+
+- Background music is not playing
+
 ## [0.4.4] - 2025/5/9
 
 ### Fixed